studio/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-orchestrator-studio"
version = "0.1.0"
description = "Visual design-time tool for creating and editing Agent-Orchestrator profiles"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "fastapi>=0.100",
    "uvicorn>=0.20",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "python-multipart>=0.0.6",
    "click>=8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
]

[project.scripts]
studio = "studio.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["studio*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
